+++ src/Rhisis.Cluster/Client/IClusterClient.cs
@@ -1,0 +1,12 @@
+﻿using Ether.Network.Common;
+
+namespace Rhisis.Cluster.Client
+{
+    public interface IClusterClient : INetUser
+    {
+        /// <summary>
+        /// Gets the ID assigned to this session.
+        /// </summary>
+        uint SessionId { get; }
+    }
+}