+++ src/Rhisis.Cluster/ClusterServer.cs
@@ -1,14 +1,15 @@
 ﻿using Ether.Network.Packets;
 using Ether.Network.Server;
+using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
-using Rhisis.Cluster.ISC;
+using Microsoft.Extensions.Options;
+using Rhisis.Cluster.Client;
+using Rhisis.Cluster.Packets;
+using Rhisis.Core.Handlers;
 using Rhisis.Core.Structures.Configuration;
 using Rhisis.Network;
 using Rhisis.Network.ISC.Structures;
-using Rhisis.Network.Packets;
 using System;
-using System.Collections.Generic;
-using System.Linq;
 
 namespace Rhisis.Cluster
 {
@@ -16,21 +17,7 @@ namespace Rhisis.Cluster
     {
         private readonly ILogger<ClusterServer> _logger;
         private readonly ClusterConfiguration _clusterConfiguration;
-
-        /// <summary>
-        /// Gets the ISC client.
-        /// </summary>
-        public static ISCClient InterClient { get; private set; }
-
-        /// <summary>
-        /// Gets the list of the connected world servers of this cluster.
-        /// </summary>
-        public static IReadOnlyCollection<WorldServerInfo> WorldServers => InterClient.WorldServers as IReadOnlyCollection<WorldServerInfo>;
-
-        /// <summary>
-        /// Gets the cluster server's configuration.
-        /// </summary>
-        public ClusterConfiguration ClusterConfiguration { get; private set; }
+        private readonly IServiceProvider _serviceProvider;
 
         /// <inheritdoc />
         protected override IPacketProcessor PacketProcessor { get; } = new FlyffPacketProcessor();
@@ -38,15 +25,20 @@ namespace Rhisis.Cluster
         /// <summary>
         /// Creates a new <see cref="ClusterServer"/> instance.
         /// </summary>
-        public ClusterServer(ILogger<ClusterServer> logger, ClusterConfiguration clusterConfiguration)
+        /// <param name="logger">Logger.</param>
+        /// <param name="clusterConfiguration">Cluster Server configuration.</param>
+        /// <param name="serviceProvider">Service provider.</param>
+        public ClusterServer(ILogger<ClusterServer> logger, IOptions<ClusterConfiguration> clusterConfiguration, IServiceProvider serviceProvider)
         {
             this._logger = logger;
-            this._clusterConfiguration = clusterConfiguration;
+            this._clusterConfiguration = clusterConfiguration.Value;
+            this._serviceProvider = serviceProvider;
             this.Configuration.Host = this._clusterConfiguration.Host;
             this.Configuration.Port = this._clusterConfiguration.Port;
             this.Configuration.MaximumNumberOfConnections = 1000;
             this.Configuration.Backlog = 100;
             this.Configuration.BufferSize = 4096;
+            this.Configuration.Blocking = false;
 
             this._logger.LogTrace("Host: {0}, Port: {1}, MaxNumberOfConnections: {2}, Backlog: {3}, BufferSize: {4}",
                 this.Configuration.Host,
@@ -59,13 +51,9 @@ namespace Rhisis.Cluster
         /// <inheritdoc />
         protected override void Initialize()
         {
-            this._logger.LogInformation("Connection to ISC server on {0}:{1}...", this._clusterConfiguration.ISC.Host, this._clusterConfiguration.ISC.Port);
-            InterClient = new ISCClient(this._clusterConfiguration);
-            InterClient.Connect();
-
             //TODO: Implement this log inside OnStarted method when will be available.
             this._logger.LogInformation("'{0}' cluster server is started and listen on {1}:{2}.", 
-                InterClient.ClusterConfiguration.Name, this.Configuration.Host, this.Configuration.Port);
+                this._clusterConfiguration.Name, this.Configuration.Host, this.Configuration.Port);
         }
 
         /// <inheritdoc />
@@ -73,7 +61,10 @@ namespace Rhisis.Cluster
         {
             this._logger.LogInformation("New client connected from {0}.", client.RemoteEndPoint);
 
-            CommonPacketFactory.SendWelcome(client, client.SessionId);
+            client.Initialize(this,
+                this._serviceProvider.GetRequiredService<ILogger<ClusterClient>>(),
+                this._serviceProvider.GetRequiredService<IHandlerInvoker>(),
+                this._serviceProvider.GetRequiredService<IClusterPacketFactory>());
         }
 
         /// <inheritdoc />
@@ -83,12 +74,9 @@ namespace Rhisis.Cluster
         }
 
         /// <inheritdoc />
-        protected override void OnError(Exception exception)
-        {
-            this._logger.LogInformation($"Socket error: {exception.Message}");
-        }
+        protected override void OnError(Exception exception) => this._logger.LogInformation($"Socket error: {exception.Message}");
 
         /// <inheritdoc />
-        public WorldServerInfo GetWorldServerById(int id) => WorldServers.FirstOrDefault(x => x.Id == id);
+        public WorldServerInfo GetWorldServerById(int id) => null;
     }
 }