+++ src/Rhisis.Cluster/Packets/ClusterPacketFactory.cs
@@ -1,38 +1,32 @@
-﻿using Ether.Network.Common;
-using Rhisis.Network;
-using Rhisis.Network.Packets;
+﻿using Rhisis.Network.Packets;
 using Rhisis.Database.Entities;
 using System.Collections.Generic;
-using System.Linq;
+using Rhisis.Network;
 using Rhisis.Core.IO;
-using Rhisis.Core.Structures.Game;
+using System.Linq;
+using Rhisis.Cluster.Client;
 
 namespace Rhisis.Cluster.Packets
 {
-    public static class ClusterPacketFactory
+    /// <summary>
+    /// Cluster packet factory.
+    /// </summary>
+    public class ClusterPacketFactory : IClusterPacketFactory
     {
-        /// <summary>
-        /// Send welcome packet to new client.
-        /// </summary>
-        /// <param name="client"></param>
-        /// <param name="sessionId"></param>
-        public static void SendWelcome(NetUser client, uint sessionId)
+        /// <inheritdoc />
+        public void SendWelcome(IClusterClient client)
         {
             using (var packet = new FFPacket())
             {
                 packet.WriteHeader(PacketType.WELCOME);
-                packet.Write(sessionId);
+                packet.Write(client.SessionId);
 
                 client.Send(packet);
             }
         }
 
-        /// <summary>
-        /// Send pong packet to client.
-        /// </summary>
-        /// <param name="client"></param>
-        /// <param name="time"></param>
-        public static void SendPong(NetUser client, int time)
+        /// <inheritdoc />
+        public void SendPong(IClusterClient client, int time)
         {
             using (var packet = new FFPacket())
             {
@@ -43,29 +37,70 @@ namespace Rhisis.Cluster.Packets
             }
         }
 
-        /// <summary>
-        /// Send an error packet to the client
-        /// </summary>
-        /// <param name="client"></param>
-        /// <param name="error"></param>
-        public static void SendError(NetUser client, ErrorType error)
+        /// <inheritdoc />
+        public void SendQueryTickCount(IClusterClient client, uint time)
         {
             using (var packet = new FFPacket())
             {
-                packet.WriteHeader(PacketType.ERROR);
-                packet.Write((int)error);
+                packet.WriteHeader(PacketType.QUERYTICKCOUNT);
+
+                packet.Write(time);
+                packet.Write(Time.GetElapsedTime());
 
                 client.Send(packet);
             }
         }
 
-        /// <summary>
-        /// Sends the client's player list.
-        /// </summary>
-        /// <param name="client"></param>
-        /// <param name="authenticationKey"></param>
-        /// <param name="characters"></param>
-        public static void SendPlayerList(NetUser client, int authenticationKey, IEnumerable<DbCharacter> characters)
+        /// <inheritdoc />
+        public void SendClusterError(IClusterClient client, ErrorType errorType)
+        {
+            using (var packet = new FFPacket())
+            {
+                packet.WriteHeader(PacketType.ERROR);
+                packet.Write((int)errorType);
+
+                client.Send(packet);
+            }
+        }
+
+        /// <inheritdoc />
+        public void SendJoinWorld(IClusterClient client)
+        {
+            using (var packet = new FFPacket())
+            {
+                packet.WriteHeader(PacketType.PRE_JOIN);
+
+                client.Send(packet);
+            }
+        }
+
+        /// <inheritdoc />
+        public void SendLoginNumPad(IClusterClient client, int loginProtectValue)
+        {
+            using (var packet = new FFPacket())
+            {
+                packet.WriteHeader(PacketType.LOGIN_PROTECT_NUMPAD);
+                packet.Write(loginProtectValue);
+
+                client.Send(packet);
+            }
+        }
+
+        /// <inheritdoc />
+        public void SendLoginProtect(IClusterClient client, int loginProtectValue)
+        {
+            using (var packet = new FFPacket())
+            {
+                packet.WriteHeader(PacketType.LOGIN_PROTECT_CERT);
+                packet.Write(0);
+                packet.Write(loginProtectValue);
+
+                client.Send(packet);
+            }
+        }
+
+        /// <inheritdoc />
+        public void SendPlayerList(IClusterClient client, int authenticationKey, IEnumerable<DbCharacter> characters)
         {
             using (var packet = new FFPacket())
             {
@@ -116,77 +151,13 @@ namespace Rhisis.Cluster.Packets
             }
         }
 
-        /// <summary>
-        /// Send the selected world server address.
-        /// </summary>
-        /// <param name="client"></param>
-        /// <param name="address"></param>
-        public static void SendWorldAddress(NetUser client, string address)
+        /// <inheritdoc />
+        public void SendWorldAddress(IClusterClient client, string address)
         {
             using (var packet = new FFPacket())
             {
                 packet.WriteHeader(PacketType.CACHE_ADDR);
                 packet.Write(address);
-
-                client.Send(packet);
-            }
-        }
-
-        /// <summary>
-        /// Send the Login num pad if the 2nd password option is enabled.
-        /// </summary>
-        /// <param name="client"></param>
-        /// <param name="loginProtectValue"></param>
-        public static void SendLoginNumPad(NetUser client, int loginProtectValue)
-        {
-            using (var packet = new FFPacket())
-            {
-                packet.WriteHeader(PacketType.LOGIN_PROTECT_NUMPAD);
-                packet.Write(loginProtectValue);
-
-                client.Send(packet);
-            }
-        }
-
-        /// <summary>
-        /// Send the login protect data.
-        /// </summary>
-        /// <param name="client"></param>
-        /// <param name="loginProtectValue"></param>
-        public static void SendLoginProtect(NetUser client, int loginProtectValue)
-        {
-            using (var packet = new FFPacket())
-            {
-                packet.WriteHeader(PacketType.LOGIN_PROTECT_CERT);
-                packet.Write(0);
-                packet.Write(loginProtectValue);
-
-                client.Send(packet);
-            }
-        }
-
-        /// <summary>
-        /// Sends a request telling the player can join the world.
-        /// </summary>
-        /// <param name="client"></param>
-        public static void SendJoinWorld(NetUser client)
-        {
-            using (var packet = new FFPacket())
-            {
-                packet.WriteHeader(PacketType.PRE_JOIN);
-
-                client.Send(packet);
-            }
-        }
-
-        public static void SendQueryTickCount(NetUser client, uint time)
-        {
-            using (var packet = new FFPacket())
-            {
-                packet.WriteHeader(PacketType.QUERYTICKCOUNT);
-
-                packet.Write(time);
-                packet.Write(Time.GetElapsedTime());
 
                 client.Send(packet);
             }