+++ src/Rhisis.Cluster/Program.cs
@@ -1,16 +1,75 @@
-﻿using Rhisis.Core;
+﻿using Ether.Network.Packets;
+using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
+using Microsoft.Extensions.Logging;
+using NLog.Extensions.Logging;
+using Rhisis.Cluster.Packets;
+using Rhisis.Core.Handlers;
+using Rhisis.Core.Structures.Configuration;
+using Rhisis.Database;
+using Rhisis.Network.Packets;
+using System.Globalization;
+using System.IO;
+using System.Text;
+using System.Threading.Tasks;
 
 namespace Rhisis.Cluster
 {
     public static class Program
     {
-        private static void Main()
+        private static async Task Main()
         {
-            ConsoleAppBootstrapper.CreateApp()
-                .SetConsoleTitle("Rhisis - Cluster Server")
-                .SetCulture("en-US")
-                .UseStartup<ClusterServerStartup>()
-                .Run();
+            const string culture = "en-US";
+            const string clusterConfigurationPath = "config/cluster.json";
+            const string databaseConfigurationPath = "config/database.json";
+
+            Encoding.RegisterProvider(CodePagesEncodingProvider.Instance);
+            CultureInfo.CurrentCulture = new CultureInfo(culture);
+            CultureInfo.CurrentUICulture = new CultureInfo(culture);
+            CultureInfo.DefaultThreadCurrentCulture = new CultureInfo(culture);
+            CultureInfo.DefaultThreadCurrentUICulture = new CultureInfo(culture);
+
+            var host = new HostBuilder()
+                .ConfigureAppConfiguration((hostContext, configApp) =>
+                {
+                    configApp.SetBasePath(Directory.GetCurrentDirectory());
+                    configApp.AddJsonFile(clusterConfigurationPath, optional: false);
+                    configApp.AddJsonFile(databaseConfigurationPath, optional: false);
+                })
+                .ConfigureServices((hostContext, services) =>
+                {
+                    services.AddOptions();
+                    services.Configure<ClusterConfiguration>(hostContext.Configuration.GetSection("clusterServer"));
+                    services.Configure<ISCConfiguration>(hostContext.Configuration.GetSection("isc"));
+                    services.RegisterDatabaseServices(hostContext.Configuration.Get<DatabaseConfiguration>());
+
+                    services.AddHandlers();
+
+                    services.AddSingleton<IClusterServer, ClusterServer>();
+                    services.AddSingleton<IClusterPacketFactory, ClusterPacketFactory>();
+                    services.AddSingleton<IHostedService, ClusterServerService>();
+                })
+                .ConfigureLogging(builder =>
+                {
+                    builder.AddFilter("Microsoft", LogLevel.Warning);
+                    builder.SetMinimumLevel(LogLevel.Trace);
+                    builder.AddNLog(new NLogProviderOptions
+                    {
+                        CaptureMessageTemplates = true,
+                        CaptureMessageProperties = true
+                    });
+                })
+                .UseConsoleLifetime()
+                .Build();
+
+            await host
+                .AddHandlerParameterTransformer<INetPacketStream, IPacketDeserializer>((source, dest) =>
+                {
+                    dest?.Deserialize(source);
+                    return dest;
+                })
+                .RunAsync();
         }
     }
 }