+++ src/Rhisis.Login/LoginServer.cs
@@ -5,6 +5,7 @@ using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using Rhisis.Core.Handlers;
 using Rhisis.Core.Structures.Configuration;
+using Rhisis.Login.Packets;
 using Rhisis.Network;
 using Rhisis.Network.ISC.Structures;
 using System;
@@ -63,7 +64,8 @@ namespace Rhisis.Login
 
             client.Initialize(this, 
                 this._serviceProvider.GetRequiredService<ILogger<LoginClient>>(), 
-                this._serviceProvider.GetRequiredService<IHandlerInvoker>());
+                this._serviceProvider.GetRequiredService<IHandlerInvoker>(),
+                this._serviceProvider.GetRequiredService<ILoginPacketFactory>());
         }
 
         /// <inheritdoc />