+++ src/Rhisis.Login/Program.cs
@@ -66,6 +66,7 @@ namespace Rhisis.Login
                         CaptureMessageProperties = true
                     });
                 })
+                .UseConsoleLifetime()
                 .Build();
 
             await host